Provide the fast-path implementation for keyedNode
The diffing algorithm has some simple heuristics that work well when
the list is mostly the same. It can detect insertion, deletion, and
moves of individual nodes. It will get confused when two of these
things are side-by-side though.

In the confusing cases, it should fall back to an alternate approach.
It is not done yet, but the fallback will be able to handle arbitrary
reordering. For example, maybe a list is sorted by date but the user
decides they want it sorted by popularity.