Add implementation for arbitrary reordering of keyed nodes
The simple heuristics cover addition, removal, and moves as long as
these operations are not side-by-side. If we run into something that is
not handled by this code, we fall back on the algorithm for arbitrary
reordering.

Before, applying a p-reorder patch meant we do a remove/update phase
followed by an insertion phase. This ensured that patches were applied
while detached from the screen and the order of inserts is not
disrupted by things that haven’t been removed yet.

This change adds another phase. We start by inserting any child nodes
that are going after all the existing nodes. These inserts will not be
disrupted by removals or insertions, they are just last. Doing these
additions first means we can skip any related removeChild operations.

This phase optimizes the existing heuristics when there are just new
children at the end of the list, but it also works for arbitrary
reordering. In that case we fall back and do everything this way. The
wager is that by using appendChild into a document fragment, we end up
being quicker than trying to do some clever in place moves.

This approach is nice because it adds very little new code, and it
works with the existing heuristics so we can do half-and-half if the
list is nice at first.

Also fix a bug with incrementing index at the wrong time in
diffKeyedChildren that was causing silly diffs.