Introduce `freeze` function for server-side rendering
The freeze function does two major things.

First, it produces an HTML string that matches the virtual-dom
structure. That HTML string will have all the attributes and styles,
but none of the event handlers.

Second, it strips the virtual-dom of all details pertaining event
handlers.

When rendering the HTML on the server, we will produce both the HTML
string an the “frozen” virtual-dom value. These correspond one-to-one,
so when you actually start the Elm code, you can start with a diff
instead of a render. This means the first update from Elm should just
be adding back event handlers!