Add freeze/thaw as alternatives to embed/fullscreen
To make server-side rendering work, it makes sense for freeze and thaw
to be integrated into the normal initialization of Elm programs. So
this will mean that you can say things like this:

    var html = Elm.MyThing.freeze(‘my-thing’);
    document.body.innerHTML = html;
    var app = Elm.MyThing.thaw(‘my-thing’);

This means you can break these three things up temporally, so you can
call freeze on the server and thaw on the client.

As a side-factor, I am adding a way to define headless programs in
core, and in that case, it ONLY makes sense to say
Elm.MyThing.worker(). So to avoid all the initialization paths being
available for all programs (and to avoid having to define freeze/thaw
in core) it was necessary to do some refactoring. Now all the
virtual-dom specific logic really lives in virtual-dom. Much nicer IMO.