Fix #32, fix subtle reorder bug
The old implementation tried to special case when there was only 1
insert at the end. The problem was that this meant that node was NOT
removed from the parent node at the right time. The new version just
puts everything in a fragment with no special case. This ensures that
all end nodes get removed at the right time.