Fix #55, avoid crash when requestAnimationFrame is not available
As I understand #55, users on IE9 and certain Android devices are seeing runtime errors because `requestAnimationFrame` is not found, and the fallback is not working.

The root cause is that [`updateIfNeeded`](https://github.com/elm-lang/virtual-dom/blob/master/src/Native/VirtualDom.js#L1549-L1574) calls itself to make sure no frames are dropped. When `rAF` does not have any delay at all, this causes an infinite loop.

This is the minimal fix for that particular problem. It will work with `elm-repl`.

It may be worthwhile to do something like #74 which uses some of the old names for `requestAnimationFrame` as well, but additional testing is needed to figure out how to get them without having `window` cause crashes in `elm-repl`.

@witoldsz and others, can you confirm that this patch also resolves the problem?