Check in an initial draft of a freezer
I experimented with a couple details in coming to this design:

1. I tried out a “string builder” implementation. That means I passed
an array around and would use builder.join(‘’) to combine everything.
This is actually not faster in Chrome, which I assume means V8 in node
will behave the same way.

2. By going with a string append approach, it becomes somewhat harder
to “stream” chunks as they are ready, but that can be done later.

3. I tried each implementation with prepack to see if that had any
influence. It has a speedup in the range of 5% for the tests I ran, and
that seemed to happen with both the builder and append strategies. So
not huge, but not bad.

4. I experimented with the fastest way to escape text. One replace with
a preallocated regex and lookup table was best.