Fix #124, stop producing unexpected names with `download`
The goal of the elm-lang/html library is to give you direct access to
HTML without any major editorial decisions. For example the <br> tag
never has attributes or children, but we allow them so that every tag
has a consistent API in this library. The changes here are in that
spirit.