Switch to simpler custom event handlers / default to synchronous
This avoids bringing in the (VirtualDom.Handler msg) type along with
VirtualDom.Timed. It also avoids confronting folks with the
capture/bubble distinction that is pointless to know about in 99.99% of
cases.

This also switches everything to synchronous rendering. This avoids
race conditions by default, but folks can opt into that world if they
need to by going to the virtual-dom library and learning about all the
details there.