Disallow a couple XSS attack vectors
Fix for https://github.com/elm-lang/html/issues/56

Ideally I would like to benchmark this pretty carefully to see what the
impact is. (Probably not noticeable.) But given that folks are
promoting this as a feature rather than a bug, it seems worthwhile to
prioritize the fix without thorough benchmarking and see if people
report any performance problems due to it during the testing phase.