ctidying and misc improvements
During this commit I was trying to track down a nasty bug in my runtime.
The bug is caused by a call to updateProcessState from within the
function passed to updateProcessState.
The result of the inner call to updateProcessState was then overwritten by
the result of the out call to updateProcessState.

Nasty!