tidy scheduler
We can do away with rawSetReceiver which was a bit of a cludge
introduced to work around the recursive dependancies between
`router`, `selfProcess` and `receiver` in
`Platform.instantiateEffectManager`. See https://github.com/elm/core/blob/f88d6a6de98802f5cc2d99dc1a0c2734cc5bdd7b/src/Elm/Kernel/Platform.js#L116-L146
for how the offical runtime manages it: using an undefined value
which is then set to the correct value later. Obviously that
approach is not possible within the elm type system. I do think
that this new approach is very nice - it stats to make clear which
bits of a process are fixed and which can change.