define List in elm and minimise kernel code
We keep kernel code for converting to and from js arrays (the
compiler may call these kernel functions) and the Nil and Cons
functions.

The Nil and Cons functions remain because a) code in elm/*
packages (e.g. elm/json) use these functions and b) because some
kernel code that runs on page load is placed above the js
generated for elm/core:List and so cannot use the elm definitions
(they will be undefined).