revisist Time for misc improvements
There was nasty memory leak in first draft. It also relied on some node
specific APIs.

This new versions does most of its work inside the onSubUpdate callback. I
think  it is fine for impure stuff to happen here provided is it the "invisible
to the user" kind of impure. For example, it is okay to set/clear timeouts from
the callback. More thought needed though.