donecallback is an impure function
This is what I have been working towards for a while now. I have tried to do
get proper use of impure functions in Task and Scheduler for a while now and
have tried to make this change a couple of times.

I failed previously because I tried to change everything in one go. I succeeded
this time because I went about it in little steps. I do really like the impure
action abstraction.