fix process killing
there was a bug whereby processes killed before they were enqueued were not
killed because the id couldn't be found in the abort actions map. Now we have a
WeakSet of process ids. Any killed id is added to this set and the scheduler
will not step any of these processes.