make Platform.Task an opaque type alias
The compiler hard codes the path of `Platform.Task` so we cannot
make it a type alias :(. What we can do is make `Platform.Task` a
stub task and use type fudging to treat it as an alias.

To make the aliasing work we have to give `RawTask.Task` an extra
type parameter. Given the options I think this is the smallest fudge
and most maintainable option.