elm spa example works!
the browser package is far from complete and a lot of functions (for
example Browser.Navigation.back) will throw runtime errors because
the kernel code is not hooked up correctly. However, completing the
browser package is just a matter of work: I am confident that all
the parts are there ready to fall into place.